+++ Expanded Cheat Menu/Code/Settings.lua
@@ -91,6 +91,8 @@ ChoGGi.Defaults = {
 	SkipModHelpPage = true,
 	-- stops panel from shrinking
 	StopSelectionPanelResize = false,
+	-- remove any objects above the height limit (or game will delete save and crash)
+	RemoveHeightLimitObjs = true
 	-- change rollovers from 450 to 600
 	WiderRollovers = 600,